Update README.md
Scharr is also used to detect the second derivatives of an image in horizontal and vertical directions. The Scharr operation can be performed on an image using the method scharr(). The Scharr operator gives a higher weight to neighboring pixels, which is more accurate than the Sobel operator.
The main idea of ​​the Scharr operator is to increase the difference between pixel values ​​by enlarging the weight coefficients in the template.  It also calculates the image difference in the x or y direction.