Update README.md
Gray level slicing displays high values in the specific region of an image and low values in other regions  by ignoring background.  It manipulates a group of  intensity  levels in  an image up  to a specific range  by  diminishing rest or  by  leaving them alone.   This transformation is applicable  in medical images and satellite images such as  X-ray flaws,  CT scans.