Update README.md
Laplacian Operator is also a derivative operator which is used to find edges in an image. It is a second order derivative mask.  It measures the rate at which first derivative changes in a single pass. Laplacian edge detection uses one kernel and contains negative values in a cross pattern. Laplacian edge detector is sensitive to noise.